pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multimodulon"
version = "0.1.0"
description = "A Python package for analyzing multi-species RNA-seq expression profiles"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "MultiModulon Development Team", email = "dev@multimodulon.org" }
]
dependencies = [
    "pandas>=1.3.0",
    "numpy>=1.20.0",
    "biopython>=1.79",
    "tqdm>=4.62.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.900",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["multimodulon*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true